git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@7285 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/ports/STM32/LLD/TIMv1/pwm_lld.c
@@ -424,7 +424,7 @@ void pwm_lld_init(void) {
  */
 void pwm_lld_start(PWMDriver *pwmp) {
   uint32_t psc;
-  uint16_t ccer;
+  uint32_t ccer;
 
   if (pwmp->state == PWM_STOP) {
     /* Clock activation and timer reset.*/
